+++ src/Vortice.Direct3D12/RaytracingGeometryAabbsDescription.cs
@@ -13,7 +13,7 @@ namespace Vortice.Direct3D12
         /// </summary>
         /// <param name="aabbCount">The number of AABBs pointed to in the contiguous array at AABBs.</param>
         /// <param name="aabbs">the GPU memory location where an array of AABB descriptions is to be found, including the data stride between AABBs. The address and stride must each be aligned to 8 bytes, defined as The address must be aligned to 16 bytes, defined as <see cref="D3D12.RaytracingAABBByteAlignment"/>. The stride can be 0.</param>
-        public RaytracingGeometryAabbsDescription(long aabbCount, GpuVirtualAddressAndStride aabbs)
+        public RaytracingGeometryAabbsDescription(ulong aabbCount, GpuVirtualAddressAndStride aabbs)
         {
             AABBCount = aabbCount;
             AABBs = aabbs;